[Fix] email result 컴포넌트 분리
