+++ front/components/account/emailResult/Result.style.ts
